+++ lib/ast_parser.ex
@@ -30,6 +30,7 @@ defmodule ExcellentMigrations.AstParser do
       detect_column_renamed(code_part) ++
       detect_column_added_with_default(code_part) ++
       detect_column_volatile_default(code_part) ++
+      detect_column_type_changed(code_part) ++
       detect_column_reference_added(code_part) ++
       detect_not_null_added(code_part) ++
       detect_check_constraint(code_part) ++
@@ -129,15 +130,22 @@ defmodule ExcellentMigrations.AstParser do
 
   defp detect_column_volatile_default(_), do: []
 
-  defp detect_column_reference_added(
-         {:modify, location, [_, {:references, _, _}, [from: {:references, _, _}]]}
-       ) do
-    [{:column_reference_added, Keyword.get(location, :line)}]
+  defp detect_column_type_changed({:modify, location, _}) do
+    [{:column_type_changed, Keyword.get(location, :line)}]
   end
 
-  defp detect_column_reference_added(
-         {fun_name, location, [_, {:references, _, [_column, options]}]}
-       )
+  defp detect_column_type_changed(_), do: []
+
+  defp detect_column_reference_added({fun_name, location, [_, {:references, _, [_, options]}]})
+       when fun_name in [:add, :modify] do
+    if Keyword.get(options, :validate) == false do
+      []
+    else
+      [{:column_reference_added, Keyword.get(location, :line)}]
+    end
+  end
+
+  defp detect_column_reference_added({fun_name, location, [_, {:references, _, [_, options]}, _]})
        when fun_name in [:add, :modify] do
     if Keyword.get(options, :validate) == false do
       []
@@ -151,8 +159,9 @@ defmodule ExcellentMigrations.AstParser do
     [{:column_reference_added, Keyword.get(location, :line)}]
   end
 
-  defp detect_column_reference_added({:modify, location, _}) do
-    [{:column_type_changed, Keyword.get(location, :line)}]
+  defp detect_column_reference_added({fun_name, location, [_, {:references, _, _}, _]})
+       when fun_name in [:add, :modify] do
+    [{:column_reference_added, Keyword.get(location, :line)}]
   end
 
   defp detect_column_reference_added(_), do: []